Split include handling into separate module
